// cargodelivery-front/generated/models/DriverIn.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cargo delivery
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DriverIn
 */
export interface DriverIn {
    /**
     * 
     * @type {string}
     * @memberof DriverIn
     */
    userId: string;
    /**
     * 
     * @type {string}
     * @memberof DriverIn
     */
    carId: string;
}

/**
 * Check if a given object implements the DriverIn interface.
 */
export function instanceOfDriverIn(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "userId" in value;
    isInstance = isInstance && "carId" in value;

    return isInstance;
}

export function DriverInFromJSON(json: any): DriverIn {
    return DriverInFromJSONTyped(json, false);
}

export function DriverInFromJSONTyped(json: any, ignoreDiscriminator: boolean): DriverIn {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'userId': json['userId'],
        'carId': json['carId'],
    };
}

export function DriverInToJSON(value?: DriverIn | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'userId': value.userId,
        'carId': value.carId,
    };
}